Device Pixel Ratio Fixes, Version Bump
**Description:** Now using a double instead of float for window scale. Extracting window scale correctly from QWindow instead of abstractly from the QScreen, fixing a scaling problem with QOpenGLWidget on high DPI displays. Minor version bump.

**Testing:** SIMDIS SDK example_qtdraganddrop, which uses both the widget and window display. Swapped between monitors at 125, 175, and 100%, and all displays worked properly after this changeset. Prior to this change, the 100% monitor displayed the Window correctly, but the Widget was showing only the bottom-left quarter or so of the display, as though it was zoomed in.

+++ include/osgQOpenGL/OSGRenderer.h
@@ -36,7 +36,7 @@ class OSGQOPENGL_EXPORT OSGRenderer : public QObject
     osg::ref_ptr<osgViewer::GraphicsWindow>    m_osgWinEmb;
     osg::observer_ptr<osgViewer::ViewerBase>   m_viewer;
     osg::ref_ptr<osgViewer::Viewer>            m_ownedViewer;
-    float                                      m_windowScale {1.0f};
+    double                                     m_windowScale {1.};
     bool                                       m_continuousUpdate {true};
 
     int                                        _timerId{0};
@@ -83,9 +83,9 @@ public:
     virtual void mouseMoveEvent(QMouseEvent* event);
     virtual void wheelEvent(QWheelEvent* event);
 
-    virtual void resize(int windowWidth, int windowHeight, float windowScale);
+    virtual void resize(int windowWidth, int windowHeight, double windowScale);
 
-    void setupOSG(int windowWidth, int windowHeight, float windowScale);
+    void setupOSG(int windowWidth, int windowHeight, double windowScale);
 
     // Calls osgViewer::ViewerBase::frame(), with main thread sleeps as needed
     void renderFrame(double simulationTime = USE_REFERENCE_TIME);
